scrapers/requirements.txt
playwright==1.50.0
psycopg2-binary==2.9.10
